integrations/database/requirements.txt
ibis-framework[duckdb,postgres,snowflake,bigquery,databricks,mssql]>=9.0.0
pyarrow>=14.0.0
pandas>=2.0.0
